feature: 数据库主键修改为bigint #885

+++ src/backend/ci/core/dispatch/biz-dispatch-docker/src/main/kotlin/com/tencent/devops/dispatch/dao/PipelineDockerDebugDao.kt
@@ -58,7 +58,7 @@ class PipelineDockerDebugDao {
         imageType: String?,
         imagePublicFlag: Boolean?,
         imageRDType: ImageRDTypeEnum?
-    ): Int {
+    ): Long {
         with(TDispatchPipelineDockerDebug.T_DISPATCH_PIPELINE_DOCKER_DEBUG) {
             val now = LocalDateTime.now()
             return dslContext.insertInto(this,
@@ -257,7 +257,7 @@ class PipelineDockerDebugDao {
         }
     }
 
-    fun deleteDebug(dslContext: DSLContext, id: Int) {
+    fun deleteDebug(dslContext: DSLContext, id: Long) {
         with(TDispatchPipelineDockerDebug.T_DISPATCH_PIPELINE_DOCKER_DEBUG) {
             dslContext.deleteFrom(this)
                 .where(ID.eq(id))
